Move files within Box, from/to trash directory

In the Box context, deleting a file moves it to a special folder within your Box account: 'Trash'. As of mid-2019, Box' default policy is to retain files in Trash for 30 days.

box_delete_file(file_id)

box_restore_file(file_id)

box_delete_folder(dir_id)

box_restore_folder(dir_id)

Arguments

file_id
numeric or character, file ID at Box.

dir_id
numeric or character, folder ID at Box.

Value

box_delete_file()

Object with S3 class boxr_file_reference.

box_restore_file()

Object with S3 class boxr_file_reference.

box_delete_folder()

Object with S3 class boxr_folder_reference.

box_restore_folder()

Object with S3 class boxr_folder_reference.

Details

box_delete_file()

Move a file to Trash.

box_restore_file()

Restore a file from Trash.

box_delete_folder()

Move a folder, including contents, to Trash.

box_restore_folder()

Restore a folder, including contents, from Trash.