Invite collaboration on a Box folder

Although this function can be used in all sorts of situations, it can be particularly useful in setting up a workflow with a service-account:

If you are authenticated as a user, using box_auth(), you can invite the service account to collaborate on a folder in your user filespace. In this case, the shared folder will appear in the service-account file-space.

If you are authenticated as as the service-account using box_auth_service(), you can invite your user-account to collaborate. In this case, the shared folder will appear in your user file-space.

Once you issue an invitation to create a collaboration, you cannot change it, e.g. you cannot change the role from "viewer" to "co-owner". However, you can delete the collaboration, then issue a new invitation. To delete a collaboration, you can use the Box web-portal.

The default role, i.e. permission level, for an invitation is "viewer". Legal values for role are "editor", "viewer", "previewer", "uploader", "previewer uploader", "viewer uploader", "co-owner", "owner".

box_dir_invite(dir_id, user_id, login = NULL, role = "viewer",
  can_view_path = FALSE)

Arguments

dir_id
numeric or character, folder ID at Box.

user_id
character ID for Box account to invite, e-mail address (login) will also work.

login
character email address of account to invite, can be used instead of user_id.

role
character role of the collaborator; default is "viewer".

can_view_path
logical indicates to allow the collaborator to navigate parent-folders at Box.

Value

Invisible list() containing collaboration-information.

Details

Regardless of the scenario, to use this function, you need the dir_id of folder you want to share and the user_id of the account with which you want to share it.

While authenticated from the host account, the one that will issue the invitation, you can use box_ls() and box_setwd() to get the dir_id for the folder you want to share. If the host-account is the user-account, you can also use the web-portal to find the dir_id. If the host account is the service-account, you can use the Box content-portal to find the dir_id.

A user can find their user_id using the Box web-portal. As well, when you authenticate using boxr, the user_id is included in the login message. Thus, you can use box_auth_service() to find out the user_id for a given service-account.

See also