List files in a Box directory

Non-recursive

box_ls(dir_id = box_getwd(), limit = 100, max = Inf, fields = NULL)

Arguments

dir_id
numeric or character, folder ID at Box.

limit
integer, maximum number of entries to retrieve per query-page.

max
integer, maximum number of entries to retrieve in total.

fields
character, fields to return; the default value, NULL, will return all possible columns: modified_at, content_modified_at, name, id, type, sha1 ,size, owned_by, path_collection, description.

Value

Object with S3 class boxr_object_list.

See also