Get details of previous versions of a Box file

Box explicitly versions files; this function returns a data.frame containing information on a file's previous versions on Box. No information about the current version of the file is returned. If the version of a file is one, then NULL is returned invisibly along with a helpful message.

box_previous_versions(file_id)

Arguments

file_id
numeric or character, file ID at Box.

Value

data.frame containing information about previous versions of the file (if available).

Details

The returned data.frame contains a variable, file_version_id, which you can use with box_dl().

References

This function is a light wrapper of the box.com API versions method.

https://developers.box.com/docs/#files-view-versions-of-a-file

See also