Download/upload an R workspace from/to a Box file

Use these functions to save and load workspaces or collections of objects to or from Box. Similar to save(), save.image(), and load(): these functions operate on files at Box instead of on local files.

Usage

box_save(..., dir_id = box_getwd(), file_name = ".RData", description = NULL)

box_save_image(
  dir_id = box_getwd(),
  file_name = ".RData",
  description = NULL,
  filename
)

box_load(file_id)

Arguments

...

Objects to be saved, quoted or unquoted; passed to save().

dir_id

numeric or character, folder ID at Box.

file_name

character, if supplied, an alternate filename for the local version of the Box file.

description

character, description caption for the file.

filename

character, deprecated: use file_name instead.

file_id

numeric or character, file ID at Box.

Value

box_save(), box_save_image()

Object with S3 class boxr_file_reference.

box_load()

From load(), a character vector of the names of objects created, invisibly.

Details

box_save()

Save object(s) using save(), write to Box.

box_save_image()

Save workspace image using save.image(), write to Box.

box_load()

Read from Box, load using load().

See also