Source R code from a Box file

Note: please use this function with extreme caution, as it is possible for the contents of a Box file to change without your knowledge.

This function downloads a file from Box, then runs its contents, as R code, using source().

Usage

box_source(file_id, local = globalenv(), ...)

Arguments

file_id

numeric or character, file ID at Box.

local

TRUE, FALSE or an environment, determining where the parsed expressions are evaluated. FALSE (the default) corresponds to the user's workspace (the global environment) and TRUE to the environment from which source is called.

...

Other arguments passed to source().

Value

Object returned by source(), called for side-effect of modifying an environment.

See also