Upload or update a file

Internal functions used by box_ul() (who's use is recommended as an alternative to these functions).

Usage

box_upload_new(dir_id, file, pb = FALSE)

box_update_file(file_id, file, dir_id, pb = FALSE)

Arguments

dir_id

The box.com id for the folder that you'd like to upload to

file

A path to a file stored locally

file_id

the box.com id of the file you'd like to update

Value

The httr::httr() object returned by the api call

Details

The box.com api requires different API calls to upload a new file, and to upload a new version of a file which already exists (incrementing the version number).

box_upload_new make the API call to upload a new file. box_update_file makes the API call to update an existing file.