Obtain a data.frame of the sub-directories in a box.com folder

Takes the id of a box folder and returns a data.frame of it's subdirectories, indluding thier equivalent paths in the local directory.

dirTreeRecursive(dir_id, local_dir = getwd())

Arguments

dir_id
The box.com id for the folder that you'd like to query

local_dir
The local directory which you'd like dir_id to correspond to. If you're not interested in mapping local to hosted directories, this isn't terribly important.

Value

A data.frame describing the contents directory structure of the box.com folder corresponding to dir_id.