Obtain a Box token

box_auth() serves two purposes:

connecting to box.com accounts from boxr for the first time

connecting to previously-connected box.com accounts

In either case, it should be sufficient to run box_auth() with no parameters. If you have authenticated with boxr before, this should be all that is required. However, the first time you use boxr, the process is slightly more involved (see 'Getting Set-Up' below).

box_auth(client_id = NULL, client_secret = NULL, interactive = TRUE,
  cache = "~/.boxr-oauth", write.Renv, ...)

Arguments

client_id
character, the client id for the account to use.

client_secret
character, the client secret for the account to use.

interactive
logical, indicates that the authorization process will be interactive (requiring user input to the R console, and/or a visit to box.com).

cache
A logical value or a string. TRUE means to cache using the default cache file .httr-oauth, FALSE means don't cache, and NA means to guess using some sensible heuristics. A string means use the specified path as the cache file.

write.Renv
deprecated.

...
Other arguments passed to httr::oauth2.0_token().

Value

invisible(NULL), called for side-effects.

See httr::oauth2.0_token() for details.

Getting Set-Up

A version of this guide is in the package vignette, with some additional screenshots. To view the vignette, run vignette("boxr"), or visit this article. To use boxr for the first time, you need to enable API access for your box.com account. The process is slightly annoying. You only need to do it once - it takes around two minutes.

The next time you use boxr, you should be able to just run box_auth() (without entering anything else) to be authenticated and ready-to-go.

Create an app:

At Box Developers Center, click on 'My Apps', in the top right hand corner log in and create a new 'app' for your box.com account. This won't create an app of any description; you'll simply be granting yourself programmatic access to your files. You can call it anything you like.

Set OAuth2 Parameters:

On the next screen, you'll want to check the box for 'Content API Access Only', and enter http://localhost as your redirect_uri.

Connect boxr to your account:

Run box_auth() and pass your client_id and client_secret to the console when prompted. These strings are not' enough for someone to access your account maliciously. However, it's still a good idea to keep them safe, and out of any files or code which might be shared with others.

A browser window should open, for you to formally grant yourself access to your files at box.com.

From this point on, simply running box_auth() at the start of a session should be all that is required.

Side-effects

This function has some side effects, which make subsequent calls to box_auth() easier:

a browser window may be opened at box.com, for you to authorize to your Box app.

a token file is written, according to the value of cache. The default behaviour is to write this file to ~/.boxr-oauth.

some global options() are set for your session to manage the token.

environment variables BOX_CLIENT_ID and BOX_CLIENT_SECRET are set.

if these environment variables have changed, and you have the usethis package installed, it will copy some text to your clipboard that you can paste into your .Renviron file.

a message is printed to the console.

See also