Read an R object from a Box file

These functions are used to download a Box file, specified by file_id, then attempt to parse its contents into memory as an R object. For example, you may wish to read a Box CSV file as a data.frame.

box_read(file_id, type = NULL, version_id = NULL, version_no = NULL,
  read_fun = rio::import, fread = FALSE, ...)

box_read_csv(file_id, ...)

box_read_tsv(file_id, ...)

box_read_json(file_id, ...)

box_read_excel(file_id, ...)

Arguments

file_id
numeric or character, file ID at Box.

type
character, MIME type used to override the content type returned by the server.

version_id
character or numeric, the version_id of the file.

version_no
numeric, version of the file you'd like to download (starting at 1).

read_fun
function, used to read (parse) the content into R; default function is rio::import().

fread
logical, indicates to use function data.table::fread() to read CSV files.

...
Other arguments passed to read_fun.

Value

Object returned by function read_fun.

Details

This is a two-step process. The first is to download the contents of the file, the second is to parse those contents into an R object. The default parsing-function is rio::import().

In addition to box_read(), some specific helpers are provided:

box_read_csv()

parse a remote CSV file into a data.frame using read.csv()

box_read_tsv()

parse a remote TSV file into a data.frame using read.delim()

box_read_json()

parse a remote JSON file into a list using jsonlite::fromJSON()

box_read_excel()

parse a remote Microsoft Excel file into a data.frame using readxl::read_excel()

See also